Clean up representation a bit
